CHE-84 Reorganized applicationController
