+++ server/controllers/applicationController/index.ts
@@ -1,0 +1,19 @@
+import createApplication from './createApplication/createApplication';
+import getAggregatedUserStats from './getAggregateUserStats/getAggregateUserStats';
+import getAllApplications from './getAllApplications/getAllApplications';
+import getApplicationById from './getApplicationById/getApplicationById';
+import getStatuses from './getStatuses/getStatuses';
+import pauseNotifications from './pauseNotifications/pauseNotifications';
+import updateApplication from './updateApplication/updateApplication';
+import updateNotificationPeriod from './updateNotificationPeriod/updateNotificationPeriod';
+
+export {
+  createApplication,
+  getAggregatedUserStats,
+  getAllApplications,
+  getApplicationById,
+  getStatuses,
+  pauseNotifications,
+  updateNotificationPeriod,
+  updateApplication,
+};
